fix: All float -> decimal

+++ HandmadeProductManagementBE/HandmadeProductManagement.ModelView/CancelReasonModelViews/CreateCancelReasonDto.cs
@@ -3,6 +3,6 @@
     public class CreateCancelReasonDto
     {
         public required string Description { get; set; }
-        public required float RefundRate { get; set; }
+        public required decimal RefundRate { get; set; }
     }
 }
